persona-generator/pyproject.toml
[project]
name = "persona-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.2",
    "streamlit>=1.50.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "plotly>=5.0.0",
]
